requirements.txt
Django==4.2.16
djangorestframework==3.14
drf-yasg[validation]==1.21.8
